common: extract SFT dataset builder into common/finetune.{h,cpp}
Move common_opt_sft_dataset_init out of common.cpp into a dedicated
finetune translation unit, dropping the nlohmann include from common.cpp.
Unify the per-template assistant span scanners into one tag-driven loop
and write samples directly into the dataset instead of staging vectors.

(cherry picked from commit d2e2077a589bfef485f56ddf2541e6efffbf5952)
(cherry picked from commit 5a6b12afe3454be1fb85f109b5b33cedcdfdfa30)
