+++ common/common.h
@@ -1234,9 +1234,3 @@ struct common_prompt_checkpoint {
     void clear_tgt();
     void clear_dft();
 };
-
-ggml_opt_dataset_t common_opt_sft_dataset_init(
-        struct llama_context * ctx,
-        const std::string    & json_content,
-        int64_t                stride,
-        const std::string    & chat_template_path = "");
